@@ -1,68 +1,75 @@
 {
-    "name": "forcrowd-wealtheconomy-angularclient",
-    "description": "Experimental ideas on increasing sustainability and productivity",
-    "author": "Stichting forCrowd",
-    "version": "0.84.0",
-    "license": "MIT",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/forCrowd/WealthEconomy"
-    },
-    "dependencies": {
-        "@angular/common": "5.2.7",
-        "@angular/core": "5.2.7",
-        "@angular/forms": "5.2.7",
-        "@angular/http": "5.2.7",
-        "@angular/platform-browser": "5.2.7",
-        "@angular/platform-browser-dynamic": "5.2.7",
-        "@angular/router": "5.2.7",
-        "angular2-highcharts": "0.5.5",
-        "angular2-moment": "1.8.0",
-        "angular2-toaster": "3.0.1",
-        "angulartics2": "3.3.0",
-        "bootstrap": "3.3.7",
-        "breeze-bridge-angular": "4.0.2",
-        "breeze-client": "1.7.0",
-        "core-js": "2.5.3",
-        "datajs": "1.0.3",
-        "font-awesome": "4.7.0",
-        "highcharts": "6.0.7",
-        "intl": "1.2.5",
-        "jquery": "3.3.1",
-        "moment": "2.20.1",
-        "respond.js": "1.4.2",
-        "rxjs": "5.5.6",
-        "source-map": "0.7.2",
-        "zone.js": "0.8.20"
-    },
-    "devDependencies": {
-        "@angular/cli": "1.7.2",
-        "@angular/compiler": "5.2.7",
-        "@angular/compiler-cli": "5.2.7",
-        "@types/jasmine": "2.8.6",
-        "@types/highcharts": "5.0.18",
-        "@types/jasminewd2": "2.0.3",
-        "@types/node": "9.4.6",
-        "@types/source-map": "0.5.7",
-        "codelyzer": "4.1.0",
-        "jasmine-core": "2.99.1",
-        "jasmine-spec-reporter": "4.2.1",
-        "karma": "2.0.0",
-        "karma-chrome-launcher": "2.2.0",
-        "karma-coverage-istanbul-reporter": "1.4.1",
-        "karma-jasmine": "1.1.1",
-        "karma-jasmine-html-reporter": "0.2.2",
-        "protractor": "5.3.0",
-        "ts-node": "5.0.1",
-        "tslint": "5.9.1",
-        "typescript": "2.6.2"
-    }
+  "name": "forcrowd-wealtheconomy-angularclient",
+  "description": "Experimental ideas on increasing sustainability and productivity",
+  "author": "Stichting forCrowd",
+  "version": "0.84.0",
+  "license": "MIT",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/forCrowd/WealthEconomy"
+  },
+  "dependencies": {
+    "@angular/animations": "6.1.8",
+    "@angular/cdk": "6.4.7",
+    "@angular/common": "6.1.8",
+    "@angular/core": "6.1.8",
+    "@angular/forms": "6.1.8",
+    "@angular/http": "6.1.8",
+    "@angular/material": "6.4.7",
+    "@angular/platform-browser": "6.1.8",
+    "@angular/platform-browser-dynamic": "6.1.8",
+    "@angular/router": "6.1.8",
+    "angular2-highcharts": "0.5.5",
+    "angular2-moment": "^1.9.0",
+    "angular2-toaster": "6.1.0",
+    "angulartics2": "6.3.0",
+    "bootstrap": "3.3.7",
+    "breeze-bridge-angular": "4.0.2",
+    "breeze-client": "1.7.1",
+    "core-js": "2.5.7",
+    "datajs": "1.0.3",
+    "font-awesome": "4.7.0",
+    "highcharts": "6.1.3",
+    "intl": "^1.2.5",
+    "jquery": "3.3.1",
+    "moment": "2.22.2",
+    "respond.js": "1.4.2",
+    "rxjs": "^6.3.2",
+    "rxjs-compat": "^6.3.2",
+    "rxjs-tslint": "^0.1.5",
+    "source-map": "^0.7.3",
+    "tsutils": "^2.29.0",
+    "zone.js": "0.8.26"
+  },
+  "devDependencies": {
+    "@angular/cli": "6.2.3",
+    "@angular/compiler": "6.1.8",
+    "@angular/compiler-cli": "6.1.8",
+    "@types/highcharts": "5.0.27",
+    "@types/jasmine": "2.8.8",
+    "@angular-devkit/build-angular": "0.8.3",
+    "@types/jasminewd2": "2.0.3",
+    "@types/node": "10.10.1",
+    "@types/source-map": "0.5.7",
+    "codelyzer": "4.4.4",
+    "jasmine-core": "3.2.1",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "3.0.0",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-coverage-istanbul-reporter": "2.0.4",
+    "karma-jasmine": "1.1.2",
+    "karma-jasmine-html-reporter": "1.3.1",
+    "protractor": "5.4.1",
+    "ts-node": "7.0.1",
+    "tslint": "^5.11.0",
+    "typescript": "2.9.2"
+  }
 }
