@@ -1,7 +1,8 @@
 ﻿import { APP_INITIALIZER, NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
-import { Angulartics2GoogleAnalytics, Angulartics2Module } from "angulartics2";
 
+import { Angulartics2Module } from 'angulartics2';
+import { Angulartics2GoogleAnalytics } from 'angulartics2/ga';
 // Components
 import { MiscComponent } from "./misc.component";
 import { ProjectTesterComponent } from "./project-tester.component";
