@@ -1,6 +1,7 @@
-﻿import { Component, OnInit } from "@angular/core";
+
+import {timer as observableTimer } from 'rxjs';
+import { Component, OnInit } from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
-import { Observable } from "rxjs";
 
 import { AuthService } from "../core/core.module";
 import { Logger } from "../logger/logger.module";
@@ -43,7 +44,7 @@ export class LoginComponent implements OnInit {
     ngOnInit(): void {
 
         // Todo This timer silliness is necessary probably cos of this issue: https://github.com/angular/angular/issues/15634
-        Observable.timer(0).subscribe(() => {
+        observableTimer(0).subscribe(() => {
 
             // Error
             const error = this.activatedRoute.snapshot.params["error"];
