@@ -1,7 +1,10 @@
-﻿import { Injectable } from "@angular/core";
+
+import {of as observableOf, throwError as observableThrowError,  Observable, Subject } from 'rxjs';
+
+import {catchError, mergeMap, map} from 'rxjs/operators';
+import { Injectable } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
 import { EntityQuery, EntityState, MergeStrategy } from "../../libraries/breeze-client";
-import { Observable, Subject } from "rxjs";
 
 import { AppSettings } from "../../app-settings/app-settings";
 import { Role } from "./entities/role";
@@ -63,7 +66,7 @@ export class AuthService {
     ensureAuthenticatedUser() {
         if (this.currentUser.isAuthenticated()) {
 
-            return Observable.of(null);
+            return observableOf(null);
 
         } else {
 
@@ -73,12 +76,12 @@ export class AuthService {
                 AutoGenerated: true
             };
 
-            return this.registerInternal(registerBindingModel)
-                .mergeMap(() => {
-                    return this.getToken("", "", true, this.currentUser.SingleUseToken).map(() => {
+            return this.registerInternal(registerBindingModel).pipe(
+                mergeMap(() => {
+                    return this.getToken("", "", true, this.currentUser.SingleUseToken).pipe(map(() => {
                         this.logger.log("Guest account has created");
-                    });
-                });
+                    }));
+                }));
         }
     }
 
@@ -89,8 +92,8 @@ export class AuthService {
             .expand("ProjectSet")
             .where("UserName", "eq", username);
 
-        return this.appEntityManager.executeQueryObservable<User>(query)
-            .map((response) => {
+        return this.appEntityManager.executeQueryObservable<User>(query).pipe(
+            map((response) => {
 
                 // If there is no result
                 if (response.results.length === 0) {
@@ -98,19 +101,19 @@ export class AuthService {
                 }
 
                 return response.results[0];
-            });
+            }));
     }
 
     init(): Observable<void> {
 
         // Get metadata from the server
-        return this.appEntityManager.getMetadata()
-            .mergeMap(() => {
+        return this.appEntityManager.getMetadata().pipe(
+            mergeMap(() => {
 
                 // Set current user
                 return this.setCurrentUser();
-            })
-            .catch((error) => {
+            }),
+            catchError((error) => {
 
                 // In case of "Server offline", set a new user to prevent further user related errors and continue
                 if (typeof error.status !== "undefined"
@@ -127,22 +130,22 @@ export class AuthService {
                     userRole.Role.Name = "Guest";
                     this.currentUser.Roles = [userRole];
 
-                    return Observable.of(null);
+                    return observableOf(null);
                 } else {
                     throw error;
                 }
-            });
+            }),);
     }
 
     login(username: string, password: string, rememberMe: boolean, singleUseToken?: string): Observable<void> {
 
-        return this.getToken(username, password, rememberMe, singleUseToken)
-            .mergeMap((): Observable<void> => {
+        return this.getToken(username, password, rememberMe, singleUseToken).pipe(
+            mergeMap((): Observable<void> => {
 
                 this.resetCurrentUser();
 
                 return this.setCurrentUser();
-            });
+            }));
     }
 
     logout(): Observable<void> {
@@ -161,21 +164,21 @@ export class AuthService {
         const restrictUsername = this.restrictUserNames.indexOf(username) > -1;
 
         if (restrictUsername) {
-            return Observable.throw("Username is already taken.");
+            return observableThrowError("Username is already taken.");
         }
 
-        return this.registerInternal(registerBindingModel)
-            .mergeMap(() => {
+        return this.registerInternal(registerBindingModel).pipe(
+            mergeMap(() => {
                 return this.getToken(registerBindingModel.UserName, registerBindingModel.Password, rememberMe);
-            });
+            }));
     }
 
     saveChanges(): Observable<void> {
 
-        return this.ensureAuthenticatedUser()
-            .mergeMap(() => {
+        return this.ensureAuthenticatedUser().pipe(
+            mergeMap(() => {
                 return this.appEntityManager.saveChangesObservable();
-            });
+            }));
     }
 
     updateCurrentUser(updatedUser: User) {
@@ -240,13 +243,13 @@ export class AuthService {
 
         if (this.appEntityManager.getEntities("Role").length > 0) {
 
-            return Observable.of(null);
+            return observableOf(null);
 
         } else {
 
             const query = EntityQuery.from("Roles");
 
-            return this.appEntityManager.executeQueryObservable(query).map(() => { });
+            return this.appEntityManager.executeQueryObservable(query).pipe(map(() => { }));
         }
     }
 
@@ -254,22 +257,22 @@ export class AuthService {
 
         const tokenData = `grant_type=password&username=${username}&password=${password}&rememberMe=${rememberMe}&singleUseToken=${singleUseToken}`;
 
-        return this.appHttpClient.post<Object>(this.tokenUrl, tokenData)
-            .map((token) => {
+        return this.appHttpClient.post<Object>(this.tokenUrl, tokenData).pipe(
+            map((token) => {
                 localStorage.setItem("token", JSON.stringify(token)); // Store the token in localStorage
-            });
+            }));
     }
 
     private registerInternal(registerBindingModel: any): Observable<void> {
 
         registerBindingModel.ClientAppUrl = window.location.origin;
 
-        return this.appHttpClient.post<User>(this.registerUrl, registerBindingModel)
-            .map((updatedUser: User) => {
+        return this.appHttpClient.post<User>(this.registerUrl, registerBindingModel).pipe(
+            map((updatedUser: User) => {
                 this.updateCurrentUser(updatedUser);
 
                 localStorage.removeItem("guestUserName");
-            });
+            }));
     }
 
     private resetCurrentUser(): void {
@@ -284,7 +287,7 @@ export class AuthService {
 
     private setCurrentUser(): Observable<void> {
 
-        return this.ensureRolesEntities().mergeMap(() => {
+        return this.ensureRolesEntities().pipe(mergeMap(() => {
 
             const tokenItem = localStorage.getItem("token");
 
@@ -294,12 +297,12 @@ export class AuthService {
 
                 this.currentUserChanged.next(this.currentUser);
 
-                return Observable.of(null);
+                return observableOf(null);
 
             } else {
 
-                return this.appHttpClient.get<User>(this.currentUserUrl)
-                    .map(currentUser => {
+                return this.appHttpClient.get<User>(this.currentUserUrl).pipe(
+                    map(currentUser => {
 
                         if (currentUser === null) {
 
@@ -316,8 +319,8 @@ export class AuthService {
                         }
 
                         this.currentUserChanged.next(this.currentUser);
-                    });
+                    }));
             }
-        });
+        }));
     }
 }
